test(runtime): share execution store contracts
